+++ modules/aegis-persistence/src/main/scala/dev/aegiskms/persistence/PostgresEventJournal.scala
@@ -99,6 +99,7 @@ object PostgresEventJournal:
         case _: KeyEvent.Activated   => "Activated"
         case _: KeyEvent.Deactivated => "Deactivated"
         case _: KeyEvent.Destroyed   => "Destroyed"
+        case _: KeyEvent.Compromised => "Compromised"
       sql"""
         INSERT INTO aegis_key_events
           (event_id, key_id, event_type, occurred_at, actor_subject, payload)